requirements.txt
pyside6
uvicorn
fastapi
starlette
python-dotenv
requests